move src/astViewer to src/language-support/ast-viewer/ast-viewer including test
