+++ extensions/ql-vscode/test/vscode-tests/no-workspace/language-support/ast-viewer/ast-viewer.test.ts
@@ -1,10 +1,13 @@
 import { readFile } from "fs-extra";
 import { load } from "js-yaml";
 
-import { AstViewer, AstItem } from "../../../src/astViewer";
+import {
+  AstViewer,
+  AstItem,
+} from "../../../../../src/language-support/ast-viewer/ast-viewer";
 import { commands, Range, Uri } from "vscode";
-import { testDisposeHandler } from "../test-dispose-handler";
-import { mockDatabaseItem } from "../utils/mocking.helpers";
+import { testDisposeHandler } from "../../../test-dispose-handler";
+import { mockDatabaseItem } from "../../../utils/mocking.helpers";
 
 describe("AstViewer", () => {
   let astRoots: AstItem[];
